Moved docs files tree fetching to fetchDocsData.ts
