+++ src/routes/+layout.svelte
@@ -1,37 +1,17 @@
 <script lang="ts">
 	import { dev } from "$app/environment";
 	import { page } from "$app/stores";
-
 	import { Footer, Navbar } from "$layout";
 	import { links, type NavbarItem } from "$data/links";
-	import type { DocsMetadata, DocsNode } from "$data/docs";
 	import { _ } from "svelte-i18n";
-
 	import "fluent-svelte/theme.css";
-
 	import Chat from "@fluentui/svg-icons/icons/chat_24_regular.svg?raw";
 	import Code from "@fluentui/svg-icons/icons/code_24_regular.svg?raw";
 	import Home from "@fluentui/svg-icons/icons/home_24_regular.svg?raw";
 	import Book from "@fluentui/svg-icons/icons/book_24_regular.svg?raw";
 	import News from "@fluentui/svg-icons/icons/news_24_regular.svg?raw";
 	// import PaintBrush from "@fluentui/svg-icons/icons/paint_brush_24_regular.svg?raw";
-
 	const { github, discord } = links;
-
-	const getDocsTree = () => {
-		const rawPages = import.meta.glob<DocsMetadata>(
-			["./docs/**/*/+page.md"], // import docs sub pages
-			{ eager: true, import: "metadata" }
-		);
-
-		return Object.entries(rawPages).map(([path, node]) => {
-			return {
-				title: node.title,
-				path: path.match(/(\/[\w-]+)+/)?.[0] ?? ""
-			} as DocsNode;
-		});
-	};
-
 	const navbarItems: NavbarItem[] = [
 		{
 			name: $_("navbar.home"),
@@ -41,7 +21,7 @@
 		{
 			name: $_("navbar.docs"),
 			path: "/docs",
-			sidebarTree: getDocsTree(),
+			sidebarTree: $page.data.docs ?? [],
 			icon: Book
 		},
 		// {
@@ -55,7 +35,6 @@
 			icon: News
 		}
 	];
-
 	const navbarButtons = [
 		{
 			label: $_("navbar.discord"),
@@ -69,12 +48,9 @@
 		}
 	];
 </script>
-
 <svelte:head>
 	<meta content="Files" name="og:site_name" />
-
 	<meta content="website" name="og:type" />
-
 	<link
 		href="/branding/logo{$page.route.id === 'themes'
 			? '-themes'
@@ -89,7 +65,6 @@
 		type="image/svg+xml"
 		media="(prefers-color-scheme: dark)"
 	/>
-
 	<meta content={$_("metadata.description")} name="description" />
 	<meta content={$_("metadata.description")} name="og:description" />
 	<meta content={$_("metadata.description")} name="twitter:description" />
@@ -98,13 +73,10 @@
 		name="keywords"
 	/>
 	<meta content="Files Community" name="author" />
-
 	<meta content="#005fb8" name="theme-color" />
-
 	<meta content="summary_large_image" name="twitter:card" />
 	<meta content="@FilesForWindows" name="twitter:site" />
 	<meta content="@FilesForWindows" name="twitter:creator" />
-
 	{#if !dev && $page.url.host === "files.community"}
 		<script type="text/javascript">
 			(function (c, l, a, r, i, t, y) {
@@ -122,11 +94,9 @@
 		</script>
 	{/if}
 </svelte:head>
-
 <Navbar buttons={navbarButtons} items={navbarItems} />
 <slot />
 <Footer />
-
 <style global lang="scss">
 	@use "src/styles/global";
 	@use "src/styles/markdown";
